% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECSFargateProvider-adaptor.R
\name{getDockerServerIp,ECSFargateProvider-method}
\alias{getDockerServerIp,ECSFargateProvider-method}
\title{Get the server public/private IPs}
\usage{
\S4method{getDockerServerIp}{ECSFargateProvider}(provider, cluster, verbose = 0L)
}
\arguments{
\item{provider}{S4 \code{CloudProvider} object. The service provider.}

\item{cluster}{S4 \code{DockerCluster} object.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
A data.frame with publicIp and privateIp columns and each row corresponds to
an element in instanceHandles.
}
\description{
Get the server public/private IPs
}
