% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECSFargateProvider-adaptor.R
\name{getDockerWorkerStatus,ECSFargateProvider-method}
\alias{getDockerWorkerStatus,ECSFargateProvider-method}
\title{Get the instance status}
\usage{
\S4method{getDockerWorkerStatus}{ECSFargateProvider}(provider, cluster, workerHandles, verbose = 0L)
}
\arguments{
\item{provider}{S4 \code{CloudProvider} object. The service provider.}

\item{cluster}{S4 \code{DockerCluster} object.}

\item{workerHandles}{Character(n). A character vector of \strong{unique} instance handles.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
A character vector with each element corresponding to a worker in workerHandles.
Each element must be one of three possible characters "initializing", "running" or "stopped"
}
\description{
Get the instance status
}
