% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECSFargateProvider-adaptor.R
\name{getServerStatus,ECSFargateProvider-method}
\alias{getServerStatus,ECSFargateProvider-method}
\title{Get the server status}
\usage{
\S4method{getServerStatus}{ECSFargateProvider}(provider, cluster, verbose)
}
\value{
Character(1)
}
\description{
Get the server status, return a character value which must be in one of three values
\code{"initializing"}, \code{"running"} or \code{"stopped"}.
}
