% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECSFargateProvider-adaptor.R
\name{killDockerWorkerContainers,ECSFargateProvider-method}
\alias{killDockerWorkerContainers,ECSFargateProvider-method}
\title{Kill the worker container}
\usage{
\S4method{killDockerWorkerContainers}{ECSFargateProvider}(provider, cluster, workerHandles, verbose = 0L)
}
\arguments{
\item{provider}{S4 \code{CloudProvider} object. The service provider.}

\item{cluster}{S4 \code{DockerCluster} object.}

\item{workerHandles}{Character(n). A character vector of \strong{unique} instance handles.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
A logical vector indicating whether the killing operation is success for each instance
}
\description{
Kill the worker container
}
