% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECSFargateProvider-adaptor.R
\name{stopDockerServer,ECSFargateProvider-method}
\alias{stopDockerServer,ECSFargateProvider-method}
\title{Stop the server container}
\usage{
\S4method{stopDockerServer}{ECSFargateProvider}(provider, cluster, verbose)
}
\arguments{
\item{provider}{S4 \code{CloudProvider} object. The service provider.}

\item{cluster}{S4 \code{DockerCluster} object.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
No return value
}
\description{
stop the server container
}
